requirements.txt
appdirs==1.4.0
beautifulsoup4==4.5.3
decorator==4.0.11
Django==1.10.5
django-bootstrap-form==3.2.1
django-extensions==1.7.6
django-test-without-migrations==0.4
django-timezone-field==2.0
django-webtest==1.8.0
djangorestframework==3.5.3
djangorestframework-jwt==1.9.0
flake8==3.3.0
ipdb==0.10.2
ipython==5.2.2
ipython-genutils==0.1.0
mccabe==0.6.1
packaging==16.8
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.13
psycopg2==2.6.2
ptyprocess==0.5.1
pycodestyle==2.3.1
pyflakes==1.5.0
Pygments==2.2.0
PyJWT==1.4.2
pyparsing==2.1.10
pytz==2016.10
simplegeneric==0.8.1
six==1.10.0
traitlets==4.3.1
waitress==1.0.2
wcwidth==0.1.7
WebOb==1.7.1
WebTest==2.0.25
